EES-5639 Moving the `CreateReleaseTests` from `ReleaseVersionServiceTests` to a new `ReleaseServiceTests`
